Move rolloutHash to the component that uses it.
